@@ -16,60 +17,60 @@
     "url": "git+https://github.com/matthewdbaldwin/VineSecret.git"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "bootstrap": "^4.5.2",
-    "jquery": "^3.5.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-redux": "^6.0.1",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "reactstrap": "^8.6.0",
-    "redux": "^4.0.1",
-    "redux-form": "^8.3.6"
+    "axios": "^1.7.2",
+    "bootstrap": "^4.6.2",
+    "jquery": "^3.7.1",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.5",
+    "react-dom": "^18.3.1",
+    "react-redux": "^8.1.3",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
+    "reactstrap": "^9.2.2",
+    "redux": "^4.2.1",
+    "redux-form": "^8.3.10"
   },
   "devDependencies": {
-    "@babel/cli": "^7.11.6",
-    "@babel/core": "^7.11.6",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-decorators": "^7.10.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.10.4",
-    "@babel/plugin-proposal-function-sent": "^7.10.4",
-    "@babel/plugin-proposal-json-strings": "^7.10.4",
-    "@babel/plugin-proposal-numeric-separator": "^7.10.4",
-    "@babel/plugin-proposal-optional-chaining": "^7.11.0",
-    "@babel/plugin-proposal-throw-expressions": "^7.10.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/eslint-parser": "^7.24.7",
+    "@babel/cli": "^7.24.7",
+    "@babel/core": "^7.24.7",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-decorators": "^7.24.7",
+    "@babel/plugin-proposal-export-namespace-from": "^7.24.7",
+    "@babel/plugin-proposal-function-sent": "^7.24.7",
+    "@babel/plugin-proposal-json-strings": "^7.24.7",
+    "@babel/plugin-proposal-numeric-separator": "^7.24.7",
+    "@babel/plugin-proposal-optional-chaining": "^7.24.7",
+    "@babel/plugin-proposal-private-methods": "^7.24.7",
+    "@babel/plugin-proposal-private-property-in-object": "^7.24.7",
+    "@babel/plugin-proposal-throw-expressions": "^7.24.7",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
     "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/preset-env": "^7.11.5",
-    "@babel/preset-react": "^7.10.4",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
-    "core-js": "^3.6.5",
-    "css-loader": "^1.0.0",
-    "eslint": "^5.15.3",
-    "eslint-plugin-react": "^7.21.0",
-    "file-loader": "^2.0.0",
-<<<<<<< HEAD
-=======
-<<<<<<< HEAD
-    "sass": "^1.77.0",
-    "postcss-loader": "^3.0.0",
-=======
->>>>>>> 5852e0babe86a6c74df179d7757ec81a2b36ff64
->>>>>>> main
-    "react-hot-loader": "^4.13.0",
-    "regenerator-runtime": "^0.13.7",
-    "style-loader": "^0.23.1",
-    "uglifyjs-webpack-plugin": "^1.1.2",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.1.4"
+    "@babel/preset-env": "^7.24.7",
+    "@babel/preset-react": "^7.24.7",
+    "babel-loader": "^9.1.3",
+    "core-js": "^3.38.0",
+    "css-loader": "^6.10.0",
+    "eslint": "^8.57.0",
+    "eslint-plugin-react": "^7.34.3",
+    "regenerator-runtime": "^0.14.1",
+    "style-loader": "^3.3.4",
+    "terser-webpack-plugin": "^5.3.10",
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.10.0"
   },
   "bugs": {
     "url": "https://github.com/matthewdbaldwin/VineSecret/issues"
   },
-  "homepage": "https://github.com/matthewdbaldwin/VineSecret#readme"
+  "homepage": "https://github.com/matthewdbaldwin/VineSecret#readme",
+  "engines": {
+    "node": ">=18"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not op_mini all"
+  ]
 }
