@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { applyMiddleware, createStore } from 'redux';
@@ -10,11 +10,13 @@ import thunk from './middleware/thunk';
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
 
-ReactDOM.render(
+const rootElement = document.getElementById('root');
+const root = createRoot(rootElement);
+
+root.render(
     <Provider store={store}>
         <Router>
             <App />
         </Router>
-    </Provider>,
-    document.getElementById('root')
+    </Provider>
 );
