@@ -1,21 +1,23 @@
 const webpack = require('webpack');
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin');
-const merge = require('webpack-merge');
+const TerserPlugin = require('terser-webpack-plugin');
+const { merge } = require('webpack-merge');
 const common = require('./webpack.common');
 
 module.exports = merge(common, {
-    entry: [
-        'core-js/stable',
-        'regenerator-runtime/runtime',
-        './index.js'
-    ],
     mode: 'production',
-    module: {
-        rules: [
-            {
-                test: /\.jsx?$/,
-                use: ['babel-loader']
-            }
+    devtool: 'source-map',
+    optimization: {
+        minimize: true,
+        minimizer: [
+            new TerserPlugin({
+                parallel: true,
+                terserOptions: {
+                    format: {
+                        comments: false
+                    }
+                },
+                extractComments: false
+            })
         ]
     },
     plugins: [
@@ -23,9 +25,6 @@ module.exports = merge(common, {
             'process.env': {
                 NODE_ENV: JSON.stringify('production')
             }
-        }),
-        new UglifyJsPlugin({
-            sourceMap: true
         })
     ]
 });
